pyproject.toml
[project]
name = "langchain-asynctools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Milansuman", email = "milansuman2016@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "pytest>=9.0.2",
]

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "langchain-litellm>=0.6.1",
    "netra-sdk>=0.1.72",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
]
